pyproject.toml
[tool.poetry]
name = "been"
version = "0.1.0"
description = "A tool to create a travel journal using photographs taken from places you have been."
authors = ["Ira Rainey"]
readme = "README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
azure-maps-search = "^2.0.0b2"
python-dotenv = "^1.0.1"
openai = "^1.57.4"
pillow = "^11.0.0"
markitdown = "^0.0.1a3"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
flake8 = "^7.1.1"
pylint = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
